rdt: separate log wrapper to its own package
This enables blockio to use the same log wrapper package
without duplicating the wrapper.
